+++ sdk/appcenter/src/main/java/com/microsoft/appcenter/Flags.java
@@ -10,11 +10,6 @@ import static com.microsoft.appcenter.utils.AppCenterLog.LOG_TAG;
 public final class Flags {
 
     /**
-     * Mask for persistence within flags.
-     */
-    private static final int PERSISTENCE_MASK = 0xFF;
-
-    /**
      * An event can be lost due to low bandwidth or disk space constraints.
      */
     public static final int PERSISTENCE_NORMAL = 0x01;
@@ -28,6 +23,11 @@ public final class Flags {
      * Default combination of flags.
      */
     public static final int DEFAULTS = PERSISTENCE_NORMAL;
+
+    /**
+     * Mask for persistence within flags.
+     */
+    private static final int PERSISTENCE_MASK = 0xFF;
 
     /**
      * Get persistence priority flag.