+++ sdk/appcenter/src/main/java/com/microsoft/appcenter/utils/storage/DatabaseManager.java
@@ -14,7 +14,6 @@ import android.support.annotation.Nullable;
 import android.support.annotation.VisibleForTesting;
 import android.text.TextUtils;
 
-import com.microsoft.appcenter.AppCenter;
 import com.microsoft.appcenter.utils.AppCenterLog;
 
 import java.io.Closeable;
@@ -38,12 +37,6 @@ public class DatabaseManager implements Closeable {
      * Primary key selection for {@link #getCursor(SQLiteQueryBuilder, String[], String[], String)}.
      */
     public static final String[] SELECT_PRIMARY_KEY = {PRIMARY_KEY};
-
-    /**
-     * Allowed multiple for maximum sizes.
-     */
-    @VisibleForTesting
-    static final int ALLOWED_SIZE_MULTIPLE = 4096;
 
     /**
      * Application context instance.
@@ -221,6 +214,7 @@ public class DatabaseManager implements Closeable {
                 } catch (SQLiteFullException e) {
 
                     /* Delete the oldest log. */
+                    AppCenterLog.debug(LOG_TAG, "Storage is full, trying to delete the oldest log that has the lowest priority which is lower or equal priority than the new log");
                     if (cursor == null) {
                         String priority = values.getAsString(priorityColumn);
                         SQLiteQueryBuilder queryBuilder = SQLiteUtils.newSQLiteQueryBuilder();
@@ -228,7 +222,9 @@ public class DatabaseManager implements Closeable {
                         cursor = getCursor(queryBuilder, SELECT_PRIMARY_KEY, new String[]{priority}, priorityColumn + " , " + PRIMARY_KEY);
                     }
                     if (cursor.moveToNext()) {
-                        delete(cursor.getLong(0));
+                        long deletedId = cursor.getLong(0);
+                        delete(deletedId);
+                        AppCenterLog.debug(LOG_TAG, "Deleted log id=" + deletedId);
                     } else {
                         throw e;
                     }
@@ -236,7 +232,7 @@ public class DatabaseManager implements Closeable {
             }
         } catch (RuntimeException e) {
             id = -1L;
-            AppCenterLog.error(AppCenter.LOG_TAG, String.format("Failed to insert values (%s) to database.", values.toString()), e);
+            AppCenterLog.error(LOG_TAG, String.format("Failed to insert values (%s) to database.", values.toString()), e);
         }
         if (cursor != null) {
             try {
@@ -268,7 +264,7 @@ public class DatabaseManager implements Closeable {
         try {
             getDatabase().execSQL(String.format("DELETE FROM " + mTable + " WHERE " + PRIMARY_KEY + " IN (%s);", TextUtils.join(", ", idList)));
         } catch (RuntimeException e) {
-            AppCenterLog.error(AppCenter.LOG_TAG, String.format("Failed to delete IDs (%s) from database.", Arrays.toString(idList.toArray())), e);
+            AppCenterLog.error(LOG_TAG, String.format("Failed to delete IDs (%s) from database.", Arrays.toString(idList.toArray())), e);
         }
     }
 
@@ -282,7 +278,7 @@ public class DatabaseManager implements Closeable {
         try {
             getDatabase().delete(mTable, key + " = ?", new String[]{String.valueOf(value)});
         } catch (RuntimeException e) {
-            AppCenterLog.error(AppCenter.LOG_TAG, String.format("Failed to delete values that match key=\"%s\" and value=\"%s\" from database.", key, value), e);
+            AppCenterLog.error(LOG_TAG, String.format("Failed to delete values that match key=\"%s\" and value=\"%s\" from database.", key, value), e);
         }
     }
 
@@ -293,7 +289,7 @@ public class DatabaseManager implements Closeable {
         try {
             getDatabase().delete(mTable, null, null);
         } catch (RuntimeException e) {
-            AppCenterLog.error(AppCenter.LOG_TAG, "Failed to clear the table.", e);
+            AppCenterLog.error(LOG_TAG, "Failed to clear the table.", e);
         }
     }
 
@@ -305,7 +301,7 @@ public class DatabaseManager implements Closeable {
         try {
             getDatabase().close();
         } catch (RuntimeException e) {
-            AppCenterLog.error(AppCenter.LOG_TAG, "Failed to close the database.", e);
+            AppCenterLog.error(LOG_TAG, "Failed to close the database.", e);
         }
     }
 
@@ -318,7 +314,7 @@ public class DatabaseManager implements Closeable {
         try {
             return DatabaseUtils.queryNumEntries(getDatabase(), mTable);
         } catch (RuntimeException e) {
-            AppCenterLog.error(AppCenter.LOG_TAG, "Failed to get row count of database.", e);
+            AppCenterLog.error(LOG_TAG, "Failed to get row count of database.", e);
             return -1;
         }
     }
@@ -384,10 +380,15 @@ public class DatabaseManager implements Closeable {
         SQLiteDatabase db = getDatabase();
         long newMaxSize = db.setMaximumSize(maxStorageSizeInBytes);
 
-        /* SQLite always use the next multiple of 4KB as maximum size. */
-        long expectedMultipleMaxSize = (long) Math.ceil((double) maxStorageSizeInBytes / (double) ALLOWED_SIZE_MULTIPLE) * ALLOWED_SIZE_MULTIPLE;
+        /* SQLite always use the next multiple of page size as maximum size. */
+        long pageSize = db.getPageSize();
+        long expectedMultipleMaxSize = maxStorageSizeInBytes / pageSize;
+        if (maxStorageSizeInBytes % pageSize != 0) {
+            expectedMultipleMaxSize++;
+        }
+        expectedMultipleMaxSize *= pageSize;
 
-        /* So to check the resize works, we need to check new max size against the next multiple of 4KB. */
+        /* So to check the resize works, we need to check new max size against the next multiple of page size. */
         if (newMaxSize != expectedMultipleMaxSize) {
             AppCenterLog.error(LOG_TAG, "Could not change maximum database size to " + maxStorageSizeInBytes + " bytes, current maximum size is " + newMaxSize + " bytes.");
             return false;
@@ -395,7 +396,7 @@ public class DatabaseManager implements Closeable {
         if (maxStorageSizeInBytes == newMaxSize) {
             AppCenterLog.info(LOG_TAG, "Changed maximum database size to " + newMaxSize + " bytes.");
         } else {
-            AppCenterLog.info(LOG_TAG, "Changed maximum database size to " + newMaxSize + " bytes (next multiple of 4KiB).");
+            AppCenterLog.info(LOG_TAG, "Changed maximum database size to " + newMaxSize + " bytes (next multiple of page size).");
         }
         return true;
     }
@@ -405,8 +406,7 @@ public class DatabaseManager implements Closeable {
      *
      * @return The maximum size of database in bytes.
      */
-    @VisibleForTesting
-    long getMaxSize() {
+    public long getMaxSize() {
         return getDatabase().getMaximumSize();
     }